# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at <a href=\"http://swagger.io\">http://swagger.io</a> or on irc.freenode.net, #swagger.  For this sample, you can use the api key \"special-key\" to test the authorization filters",
        "termsOfService": "http://helloreverb.com/terms/",
        "contact": {
            "email": "apiteam@wordnik.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:1337",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/abonnement": {
            "get": {
                "tags": [
                    "abonnement"
                ],
                "summary": "Display a listing of all abonnement.",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Abonnement"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "abonnement"
                ],
                "summary": "Store a new created abonnement in storage.",
                "operationId": "postAbonnement",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id_forfait",
                        "in": "formData",
                        "description": "id of forfait",
                        "type": "integer"
                    },
                    {
                        "name": "debut",
                        "in": "formData",
                        "description": "Time of the abonnement",
                        "type": "string",
                        "format": "date"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Abonnement"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/abonnement/{id_abonnement}": {
            "get": {
                "tags": [
                    "abonnement"
                ],
                "summary": "Display the specified abonnement.",
                "operationId": "getAbonnement",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id_abonnement",
                        "in": "path",
                        "description": "Abonnement id to get",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Abonnement"
                        }
                    },
                    "404": {
                        "description": "Abonnement does not exist"
                    }
                }
            },
            "put": {
                "tags": [
                    "abonnement"
                ],
                "summary": "Update an existing abonnement in storage.",
                "operationId": "putAbonnement",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id_forfait",
                        "in": "formData",
                        "description": "id of forfait",
                        "type": "integer"
                    },
                    {
                        "name": "debut",
                        "in": "formData",
                        "description": "Time of the abonnement",
                        "type": "string",
                        "format": "date"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Abonnement does not exist"
                    }
                }
            },
            "delete": {
                "tags": [
                    "abonnement"
                ],
                "summary": "Remove the specified abonnement from storage.",
                "operationId": "deleteAbonnement",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id_abonnement",
                        "in": "path",
                        "description": "Abonnement identifier",
                        "schema": {
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Abonnement does not exist"
                    }
                }
            }
        },
...
    },
...
}